requirements.txt
asgiref==3.11.1
Django==6.0.3
django-sesame==3.2.3
Pillow>=10.0.0
sqlparse==0.5.5
